合约下单

+++ base/src/main/java/com/black/base/model/future/UserTradeBean.kt
@@ -5,5 +5,5 @@ data class UserTradeBean(
     val orderId: String,
     val price: String,
     val quantity: String,
-    val timestamp: Int
+    val timestamp: Long
 )
